Update the multitenant migration command to use the code from cmd/db, insted of the other way around.
